service tree model moved to domain module
