+++ publisher-domain/src/main/java/nl/idgis/publisher/domain/web/tree/Group.java
@@ -1,0 +1,5 @@
+package nl.idgis.publisher.domain.web.tree;
+
+public interface Group extends Item {
+
+}
