+++ publisher-domain/src/main/java/nl/idgis/publisher/domain/web/tree/GroupNode.java
@@ -1,10 +1,10 @@
-package nl.idgis.publisher.service.manager.messages;
+package nl.idgis.publisher.domain.web.tree;
 
 import com.mysema.query.annotations.QueryProjection;
 
 public class GroupNode extends Node implements Group {
 
-	private static final long serialVersionUID = -1256215511214933273L;
+	private static final long serialVersionUID = 7986334970024624615L;
 
 	@QueryProjection
 	public GroupNode(String id, String name, String title, String abstr) {
