+++ publisher-domain/src/main/java/nl/idgis/publisher/domain/web/tree/Layer.java
@@ -1,8 +1,11 @@
-package nl.idgis.publisher.service.manager.messages;
+package nl.idgis.publisher.domain.web.tree;
+
 
 public interface Layer extends Item {
 	
 	boolean isGroup();
+	
 	GroupLayer asGroup();
+	
 	DatasetLayer asDataset();
 }
