+++ publisher-domain/src/main/java/nl/idgis/publisher/domain/web/tree/Service.java
@@ -1,10 +1,12 @@
-package nl.idgis.publisher.service.manager.messages;
+package nl.idgis.publisher.domain.web.tree;
 
 import java.util.List;
 
 public interface Service {
 
 	String getId();
+	
 	String getRootId();
+	
 	List<Layer> getLayers();
 }
