--- publisher-service/src/main/java/nl/idgis/publisher/service/manager/messages/Group.java
@@ -1,5 +1,0 @@
-package nl.idgis.publisher.service.manager.messages;
-
-public interface Group extends Item {
-
-}
